+++ src/main/java/units/coin/Coin.java
@@ -3,7 +3,7 @@ package main.java.units.coin;
 import main.java.units.Unit;
 
 public class Coin extends Unit<CoinUnits> {
-    public Coin(CoinUnits coinUnits, float amount) {
+    public Coin(CoinUnits coinUnits, double amount) {
 	    super(coinUnits, amount);
     }
 }
