+++ src/main/java/units/weight/Weight.java
@@ -3,7 +3,7 @@ package main.java.units.weight;
 import main.java.units.Unit;
 
 public class Weight extends Unit<WeightUnits> {
-	public Weight(WeightUnits weightUnits, float amount) {
+	public Weight(WeightUnits weightUnits, double amount) {
 		super(weightUnits, amount);
 	}
 }
